Split migration into classes
One file became unwieldy, I split the classes into different files.
The namespace can be deleted when the migration is finished
