+++ src/DataAccess/PaymentMigration/BoundedValue.php
@@ -1,0 +1,23 @@
+<?php
+declare( strict_types=1 );
+
+namespace WMDE\Fundraising\DonationContext\DataAccess\PaymentMigration;
+
+class BoundedValue {
+	private mixed $lowerBound;
+	private mixed $upperBound;
+
+	public function __construct( mixed $value ) {
+		$this->lowerBound = $value;
+		$this->upperBound = $value;
+	}
+
+	public function set( mixed $value ): void {
+		if ( $value < $this->lowerBound ) {
+			$this->lowerBound = $value;
+		}
+		if ( $value > $this->upperBound ) {
+			$this->upperBound = $value;
+		}
+	}
+}
